Implementacao e mapeamento alltypes e ajustes nos mapeamentos de series

+++ backend/src/main/java/com/estagio2/myhappyplace/entities/FavoriteSeries.java
@@ -27,5 +27,5 @@ public class FavoriteSeries {
     private List<User> userList = new ArrayList<User>();
 
     @Column(name = "serieId_TMDB")
-    private Long serieId;
+    private Integer serieId;
 }
